Moved getVesselCommunicationsFields() to separate file. Tweaked more styles
