+++ backoffice/src/certificate.scss
@@ -15,6 +15,10 @@
   float: left;
 }
 
+.logoContainer {
+  padding-left: 48%;
+}
+
 .mcaLogo {
   height: 90px;
   float: right;
